# docs/swagger.yaml
basePath: /api/v1
consumes:
- application/json
definitions:
...
  models.AuthorInfo:
    properties:
      avatar:
        $ref: '#/definitions/models.Photo'
      first_name:
        type: string
      full_name:
        type: string
      last_name:
        type: string
      user_id:
        type: string
    type: object
...
  models.BusinessInfo:
    properties:
      avatar:
        $ref: '#/definitions/models.Photo'
      business_id:
        type: string
      name:
        type: string
    type: object
...
  models.BusinessSearchResult:
    properties:
      address:
        type: string
      avatar:
        $ref: '#/definitions/models.Photo'
      categories:
        items:
          type: string
        type: array
      cover:
        $ref: '#/definitions/models.Photo'
      description:
        type: string
      distance:
        description: Distance in km from search point
        type: number
      id:
        type: string
      is_following:
        type: boolean
      location:
        $ref: '#/definitions/models.Location'
      name:
        type: string
      phone_number:
        type: string
      total_follow:
        type: integer
      total_views:
        type: integer
      website:
        type: string
    type: object
  models.CategoryInfo:
    properties:
      color:
        type: string
      icon:
        $ref: '#/definitions/models.Icon'
      id:
        type: string
      name:
        type: string
    type: object
...
  models.DiscoverBusiness:
    properties:
      avatar:
        $ref: '#/definitions/models.Photo'
      categories:
        items:
          type: string
        type: array
      description:
        type: string
      distance:
        description: Distance in km from search point
        type: number
      id:
        type: string
      location:
        $ref: '#/definitions/models.Location'
      name:
        type: string
      total_follow:
        type: integer
    type: object
  models.DiscoverPost:
    properties:
      created_at:
        type: string
      description:
        type: string
      distance:
        description: Distance in km from search point
        type: number
      id:
        type: string
      location:
        $ref: '#/definitions/models.Location'
      price:
        type: number
      start_date:
        description: For EVENT type
        type: string
      start_time:
        type: string
      thumbnail:
        allOf:
        - $ref: '#/definitions/models.Photo'
        description: First attachment
      title:
        type: string
      type:
        $ref: '#/definitions/models.PostType'
    type: object
  models.DiscoverResponse:
    properties:
      businesses:
        items:
          $ref: '#/definitions/models.DiscoverBusiness'
        type: array
      posts:
        items:
          $ref: '#/definitions/models.DiscoverPost'
        type: array
      total:
        type: integer
    type: object
...
  models.EventState:
    enum:
    - upcoming
    - ongoing
    - ended
    type: string
    x-enum-varnames:
    - EventStateUpcoming
    - EventStateOngoing
    - EventStateEnded
...
  models.Icon:
    properties:
      library:
        type: string
      name:
        type: string
    type: object
  models.Location:
    properties:
      country:
        type: string
      district:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      province:
        type: string
    type: object
  models.LocationInfo:
    properties:
      country:
        type: string
      district:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      neighborhood:
        type: string
      province:
        type: string
    type: object
...
  models.Photo:
    properties:
      height:
        type: integer
      mime_type:
        type: string
      name:
        type: string
      size:
        type: integer
      url:
        type: string
      width:
        type: integer
    type: object
...
  models.PostResponse:
    properties:
      attachments:
        description: Attachments
        items:
          $ref: '#/definitions/models.Photo'
        type: array
      author:
        allOf:
        - $ref: '#/definitions/models.AuthorInfo'
        description: Author info
      bookmarked_by_me:
        type: boolean
      business:
        $ref: '#/definitions/models.BusinessInfo'
      category:
        $ref: '#/definitions/models.CategoryInfo'
      contact_no:
        type: string
      created_at:
        description: Timestamps
        type: string
      currency:
        description: Sell-specific
        type: string
      description:
        type: string
      discount:
        type: number
      end_date:
        type: string
      end_time:
        type: string
      event_state:
        $ref: '#/definitions/models.EventState'
      free:
        type: boolean
      going_count:
        type: integer
      id:
        type: string
      interested_count:
        type: integer
      is_promoted:
        type: boolean
      liked_by_me:
        type: boolean
      location:
        allOf:
        - $ref: '#/definitions/models.LocationInfo'
        description: Location
      original_post:
        allOf:
        - $ref: '#/definitions/models.PostResponse'
        description: Original post (for shares)
      price:
        type: number
      sold:
        type: boolean
      start_date:
        description: Event-specific
        type: string
      start_time:
        type: string
      status:
        type: boolean
      title:
        type: string
      total_comments:
        description: Engagement
        type: integer
      total_likes:
        type: integer
      total_shares:
        type: integer
      type:
        $ref: '#/definitions/models.PostType'
      updated_at:
        type: string
      visibility:
        $ref: '#/definitions/models.PostVisibility'
    type: object
  models.PostType:
    enum:
    - FEED
    - EVENT
    - SELL
    - PULL
    type: string
    x-enum-varnames:
    - PostTypeFeed
    - PostTypeEvent
    - PostTypeSell
    - PostTypePull
  models.PostVisibility:
    enum:
    - PUBLIC
    - FRIENDS
    - PRIVATE
    type: string
    x-enum-varnames:
    - VisibilityPublic
    - VisibilityFriends
    - VisibilityPrivate
...
  models.SearchResponse:
    properties:
      businesses:
        items:
          $ref: '#/definitions/models.BusinessSearchResult'
        type: array
      posts:
        items:
          $ref: '#/definitions/models.PostResponse'
        type: array
      total:
        type: integer
      users:
        items:
          $ref: '#/definitions/models.UserSearchResult'
        type: array
    type: object
...
  models.UserSearchResult:
    properties:
      about:
        type: string
      avatar:
        $ref: '#/definitions/models.Photo'
      first_name:
        type: string
      full_name:
        type: string
      id:
        type: string
      is_followed_by:
        type: boolean
      is_following:
        description: Relationship status
        type: boolean
      last_name:
        type: string
      province:
        type: string
    type: object
...
  utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@hamsaya.app
    name: API Support
    url: https://hamsaya.app/support
  description: A production-ready Go backend for a social media mobile application
    with posts, business profiles, marketplace, real-time chat, and location services.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://hamsaya.app/terms/
  title: Hamsaya Backend API
  version: "1.0"
paths:
...
  /discover:
    get:
      consumes:
      - application/json
      description: Get posts and businesses within a radius for map view
      parameters:
      - description: Latitude
        in: query
        name: latitude
        required: true
        type: number
      - description: Longitude
        in: query
        name: longitude
        required: true
        type: number
      - description: Radius in kilometers (max 100)
        in: query
        name: radius_km
        required: true
        type: number
      - description: 'Post type filter: FEED, EVENT, SELL, PULL'
        in: query
        name: type
        type: string
      - description: Limit results (default 100, max 500)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.DiscoverResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Map-based discovery
      tags:
      - Discovery
...
  /search:
    get:
      consumes:
      - application/json
      description: Search across posts, users, and businesses
      parameters:
      - description: Search query (minimum 2 characters)
        in: query
        name: query
        required: true
        type: string
      - description: 'Search type: all, posts, users, businesses'
        enum:
        - all
        - posts
        - users
        - businesses
        in: query
        name: type
        type: string
      - description: Limit per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      - description: Latitude for location-based search
        in: query
        name: latitude
        type: number
      - description: Longitude for location-based search
        in: query
        name: longitude
        type: number
      - description: Radius in kilometers for location-based search
        in: query
        name: radius_km
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.SearchResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Global search
      tags:
      - Search
  /search/businesses:
    get:
      consumes:
      - application/json
      description: Search for businesses by name, description, or location
      parameters:
      - description: Search query
        in: query
        name: query
        required: true
        type: string
      - description: Limit per page
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      - description: Latitude for location-based search
        in: query
        name: latitude
        type: number
      - description: Longitude for location-based search
        in: query
        name: longitude
        type: number
      - description: Radius in kilometers
        in: query
        name: radius_km
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.BusinessSearchResult'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Search businesses
      tags:
      - Search
  /search/posts:
    get:
      consumes:
      - application/json
      description: Search for posts with full-text search and location filtering
      parameters:
      - description: Search query
        in: query
        name: query
        required: true
        type: string
      - description: Limit per page
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      - description: Latitude for location-based search
        in: query
        name: latitude
        type: number
      - description: Longitude for location-based search
        in: query
        name: longitude
        type: number
      - description: Radius in kilometers
        in: query
        name: radius_km
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.PostResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Search posts
      tags:
      - Search
  /search/users:
    get:
      consumes:
      - application/json
      description: Search for users by name
      parameters:
      - description: Search query
        in: query
        name: query
        required: true
        type: string
      - description: Limit per page
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.UserSearchResult'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Search users
      tags:
      - Search
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"